Refactored and Optimized the code base with basic UI element added in tablet version
